Refactor types, update dependencies, and refactor tests
- Changed `Years` property in `DuplicateNamesAndBirthdaysResponse` from `short` to `byte`
- Changed `VestedPercent` property in `TerminatedEmployeeAndBeneficiaryDataResponseDto` from `int` to `decimal`
- Updated `years` variable type in `CleanupReportService` to `byte`
- Removed unused import from `TerminatedEmployeeAndBeneficiaryReport.cs`
- Updated `currentVestedAmount` calculation in `TerminatedEmployeeAndBeneficiaryReport`
- Added `TotalService` and `ContributionService` dependencies to `TerminatedEmployeeAndBeneficiaryReportService`
- Updated `TerminatedEmployeeAndBeneficiaryReportIntegrationTests` to use new dependencies
- Changed `wVestPert` parameter type in `PrintDetails` method of `TextReportGenerator` to `decimal`
- Updated `BeneficiaryContactFaker` to use nested `ContactInfo` properties
- Updated `TerminatedEmployeeAndBeneficiaryTests` to use new dependencies
- Refactored `TerminatedEmployee` and `SingleBeneficiary` test methods to return tasks directly
- Updated `TerminatedEmployee` test method to use `EmployeeId` instead of `BadgeNumber`
- Updated `SingleBeneficiary` test method to use nested `ContactInfo` properties

+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Response/YearEnd/DuplicateNamesAndBirthdaysResponse.cs
@@ -6,7 +6,7 @@ public sealed record DuplicateNamesAndBirthdaysResponse
     public string? Name { get; set; }
     public required DateOnly DateOfBirth { get; set; }
     public required AddressResponseDto Address { get; set; }
-    public short Years { get; set; }
+    public byte Years { get; set; }
     public DateOnly HireDate { get; set; }
     public DateOnly? TerminationDate { get; set; }
     public required char Status { get; set; }
